Introduce tests for redis auto-reconfiguration
The bulk of changes refactor the existing structure to avoid code duplication
that would otherwise be caused when introducing a tests for new services
